Clean up 1st round.

+++ arch/h8300/include/asm/page.h
@@ -1,5 +1,9 @@
 #ifndef _H8300_PAGE_H
 #define _H8300_PAGE_H
+
 #include <asm-generic/page.h>
-#define MAP_NR(addr)		(((unsigned long)(addr)-PAGE_OFFSET) >> PAGE_SHIFT)
+#include <linux/types.h>
+
+#define MAP_NR(addr) (((uintptr_t)(addr)-PAGE_OFFSET) >> PAGE_SHIFT)
+
 #endif
